@@ -65,4 +65,4 @@ const nextConfig = {
   },
 }
 
-module.exports = nextConfig
+export default nextConfig
